Mock Bamboo copy repository
Also refactor BambooUpdateService
