+++ src/main/java/de/tum/in/www1/artemis/config/connector/bamboo/BambooServerConfiguration.java
@@ -1,5 +1,7 @@
 package de.tum.in.www1.artemis.config.connector.bamboo;
 
+import java.io.ByteArrayOutputStream;
+import java.io.PrintStream;
 import java.net.URL;
 
 import org.springframework.beans.factory.annotation.Value;
@@ -7,6 +9,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 
+import com.appfire.bamboo.cli.BambooClient;
+import com.appfire.common.cli.Base;
+import com.appfire.common.cli.Settings;
 import com.atlassian.bamboo.specs.util.BambooServer;
 import com.atlassian.bamboo.specs.util.SimpleUserPasswordCredentials;
 import com.atlassian.bamboo.specs.util.UserPasswordCredentials;
@@ -24,9 +29,58 @@ public class BambooServerConfiguration {
     @Value("${artemis.bamboo.url}")
     private URL BAMBOO_SERVER_URL;
 
+    @Value("${artemis.version-control.user}")
+    private String BITBUCKET_USER;
+
+    @Value("${artemis.version-control.secret}")
+    private String BITBUCKET_PASSWORD;
+
+    @Value("${artemis.version-control.url}")
+    private URL BITBUCKET_SERVER;
+
     @Bean
     public BambooServer bambooServer() {
         UserPasswordCredentials userPasswordCredentials = new SimpleUserPasswordCredentials(BAMBOO_USER, BAMBOO_PASSWORD);
         return new BambooServer(BAMBOO_SERVER_URL.toString(), userPasswordCredentials);
+    }
+
+    /**
+     * Create a BambooClient for communication with the Bamboo server.
+     *
+     * @return BambooClient instance for the Bamboo server that is defined in the environment yml files.
+     */
+    @Bean
+    @Profile("bitbucket")
+    public BambooClient bambooClient() {
+        final var bambooClient = new BambooClient(createBase());
+        // setup the Bamboo Client to use the correct username and password
+        final var args = new String[] { "-s", BAMBOO_SERVER_URL.toString(), "--user", BAMBOO_USER, "--password", BAMBOO_PASSWORD, "--targetServer", BITBUCKET_SERVER.toString(),
+                "--targetUser", BITBUCKET_USER, "--targetPassword", BITBUCKET_PASSWORD };
+
+        bambooClient.doWork(args); // only invoke this to set server address, username and password so that the following action will work
+        return bambooClient;
+    }
+
+    @Bean("bambooClient")
+    @Profile("!bitbucket")
+    public BambooClient bambooClientWithoutBitbucket() {
+        final var bambooClient = new BambooClient(createBase());
+        // setup the Bamboo Client to use the correct username and password
+        final var args = new String[] { "-s", BAMBOO_SERVER_URL.toString(), "--user", BAMBOO_USER, "--password", BAMBOO_PASSWORD };
+
+        bambooClient.doWork(args); // only invoke this to set server address, username and password so that the following action will work
+        return bambooClient;
+    }
+
+    private Base createBase() {
+        // we override the out stream to prevent unnecessary log statements in our log files
+        final var outputStream = new ByteArrayOutputStream();
+        final var printStream = new PrintStream(outputStream);
+        final var settings = new Settings();
+        settings.setOut(printStream);
+        settings.setOverrideOut(printStream);
+        settings.setDebugOut(printStream);
+        settings.setErr(printStream);
+        return new Base(settings);
     }
 }
