+++ src/main/java/de/tum/in/www1/artemis/service/connectors/BitbucketService.java
@@ -423,7 +423,12 @@ public class BitbucketService implements VersionControlService {
                     else {
                         throw e;
                     }
+
+                    // Try again if there was an exception
+                    continue;
                 }
+                // Don't try again if everything went fine
+                break;
             }
         }
         catch (HttpClientErrorException e) {
