+++ src/test/java/de/tum/in/www1/artemis/util/Verifiable.java
@@ -1,0 +1,7 @@
+package de.tum.in.www1.artemis.util;
+
+@FunctionalInterface
+public interface Verifiable {
+
+    void performVerification() throws Exception;
+}
